change prototype to use "size_t" since it is a buffer size, 	and should never be negative.

+++ src/include/radiusd.h
@@ -345,7 +345,7 @@ typedef int (*RADIUS_ESCAPE_STRING)(char *out, int outlen, const char *in);
 
 int            radius_xlat(char * out, int outlen, const char *fmt,
 			   REQUEST * request, RADIUS_ESCAPE_STRING func);
-typedef int (*RAD_XLAT_FUNC)(void *instance, REQUEST *, char *, char *, int, RADIUS_ESCAPE_STRING func);
+typedef int (*RAD_XLAT_FUNC)(void *instance, REQUEST *, char *, char *, size_t, RADIUS_ESCAPE_STRING func);
 int		xlat_register(const char *module, RAD_XLAT_FUNC func, void *instance);
 void		xlat_unregister(const char *module, RAD_XLAT_FUNC func);
